% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrapping.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrapping}
\usage{
bootstrap(X, n)
}
\arguments{
\item{X}{Data}

\item{n}{Number of Trials}
}
\value{
Table of the mean, standard error and 95% confidence intervals of the bootstrapped Alpha and Beta values.
}
\description{
Bootstrapping
}
\examples{
boostrap(data$V1,10^3)
}
