% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed_summary.R
\name{detailed_summary}
\alias{detailed_summary}
\title{Detailed Summary Statistics}
\usage{
detailed_summary(x)
}
\arguments{
\item{x}{Variable to be Analyzed}
}
\value{
Table of more summary statistics, including mean, standard deviation, minimum value, 25th percentile, median, 75th percentile, maximum value, interquartile range, upper limit and lower limit.
}
\description{
Detailed Summary Statistics
}
\examples{
detailed_summary(data$V1)
}
