% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy}
\alias{dummy}
\title{Dummy Variable}
\usage{
dummy(x, i, c)
}
\arguments{
\item{x}{Specified Variable}

\item{i}{Inequality (Describes relationship between the variable and the condition) (i.e. "=","<",">=")}

\item{c}{Condition}
}
\value{
Generates dummy variables for a variable under a specified relationship
}
\description{
Dummy Variable
}
\examples{
dummy(data$V1,"<",5) -> generates dummy variables equal to 1 if data$V1<5, 0 otherwise.
dummy(data$V1,"=","MALE") -> generates dummy variables that are equal to 1 if data$V1 is "MALE", 0 otherwise.
}
