% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MallowsCP.R
\name{mcp}
\alias{mcp}
\title{Mallow's CP}
\usage{
mcp(x, d)
}
\arguments{
\item{x}{Data in Regression Form}

\item{d}{Source of data in regression}
}
\value{
Best Regression Functional Form According to Mallow's CP
}
\description{
Mallow's CP
}
\examples{
mcp(df$wage~df$gender+df$age,df)
}
